Login Controller + login stratgies , User and superadmin models  (#1)
* fix db columns and add more repository fuctions


* added LoginController , SignUpController, LoginStrategies


---------

+++ charity-org/src/main/java/com/charity_org/demo/Models/Address.java
@@ -7,6 +7,6 @@ import jakarta.persistence.Entity;
 @Data
 public class Address extends BaseEntity {
     private String name;
-    private int parentID;
+    private long parentID;
 }
